Split nsmgr test files
